separation of logging config into separte files
