+++ src/wavealign/utility/logging/filters.py
@@ -1,0 +1,6 @@
+import logging
+
+
+class ExcludeWarningsFilter(logging.Filter):
+    def filter(self, record: logging.LogRecord) -> bool:
+        return record.levelno < logging.WARNING
