+++ src/wavealign/utility/logging/logging_variables.py
@@ -1,0 +1,2 @@
+LOGFILE_NAME = "wavealign.log"
+warning_counts = False
